BUGFIX: #46 introduced a bug in how columns were counted, and broke when a line had a "description" field.
Removed the "recalculateColumns()" method, since each "recalculation" was simply incrementing one the number of columns. We can initialize the number of columns as 1, and $this->columns++ when needed.

added a getOtherColumnsWith(), since this calculation is performed a couple of times in the class. since it was the source of the bug, better to fix it just once.